Open-Lab-01.4

+++ Open-Lab-01.4/Program.cs
@@ -9,9 +9,9 @@ namespace Open_Lab_01._4
             Console.WriteLine("Type the base of the triangle and then enter its height");
             string siteA = Console.ReadLine();
             string height = Console.ReadLine();
-            int a = Convert.ToInt32(siteA);
-            int b = Convert.ToInt32(height);
-            int c = a * b / 2;
+            double a = Convert.ToInt32(siteA);
+            double b = Convert.ToInt32(height);
+            double c = a * b / 2;
             Console.WriteLine($"The area of ​​the triangle is {c}.");
         }
     }
